Task 1:part 2 THOUGHTS
the spreadsheet is a grid but rather than creating the complete grid
when the class is instantiated I'll only create a cell when it is
populated storing the grid in a hash or hashes

so for example

spreadsheet->populate(cell => 'D2', value =>'foo')

will populate the internal variable _grid as follows

spreadsheet->_grid->{'D'}{'2'} = cell->new(value => 'foo')

Assuming the ref is not already populated with a cell instance.

To test the above logic I'll need to create a method to retrieve the
value of a given cell.

spreadsheet->getCellValue(cell => 'A1')

This will return 0 if the cell has not been instantiated yet. Otherwise
it will delegate the responsiblity of returning a value to the cell