THOUGHTS: validation
I need a method of knowing that 'D1' maps to a valid grid location on a
spreadsheet.

validation of rows is easy as its numeric ($row < spreadsheet->rows)

but that flops with characters ('b' gt 'bb')

best method is to build a valid hash of rows and cols then compare cell
to valid sets.